+++ src/sdk/modules/albedo.module.ts
@@ -34,8 +34,6 @@ export class AlbedoModule implements ModuleInterface {
       networkPassphrase?: string;
       address?: string;
       path?: string;
-      submit?: boolean;
-      submitUrl?: string;
     },
   ): Promise<{ signedTxXdr: string; signerAddress?: string }> {
     try {