+++ src/sdk/modules/hana.module.ts
@@ -70,8 +70,6 @@ export class HanaModule implements ModuleInterface {
       networkPassphrase?: string;
       address?: string;
       path?: string;
-      submit?: boolean;
-      submitUrl?: string;
     },
   ): Promise<{ signedTxXdr: string; signerAddress?: string }> {
     try {