+++ src/sdk/modules/lobstr.module.ts
@@ -38,8 +38,6 @@ export class LobstrModule implements ModuleInterface {
       networkPassphrase?: string;
       address?: string;
       path?: string;
-      submit?: boolean;
-      submitUrl?: string;
     },
   ): Promise<{ signedTxXdr: string; signerAddress?: string }> {
     if (opts?.address) {