+++ src/sdk/modules/rabet.module.ts
@@ -49,8 +49,6 @@ export class RabetModule implements ModuleInterface {
       networkPassphrase?: string;
       address?: string;
       path?: string;
-      submit?: boolean;
-      submitUrl?: string;
     },
   ): Promise<{ signedTxXdr: string; signerAddress?: string }> {
     if (