+++ src/sdk/modules/wallet-connect.module.ts
@@ -55,7 +55,8 @@ export class WalletConnectModule implements ModuleInterface {
       enableReconnect: true,
       networks: [mainnet],
       featuredWalletIds: [
-        "aef3112adf415ec870529e96b4d7b434f13961a079d1ee42c9738217d8adeb91",
+        "aef3112adf415ec870529e96b4d7b434f13961a079d1ee42c9738217d8adeb91", // Freighter
+        "76a3d548a08cf402f5c7d021f24fd2881d767084b387a5325df88bc3d4b6f21b", // Lobstr
       ],
       ...(wcParams.appKitOptions || {}),
     });