+++ src/sdk/modules/xbull.module.ts
@@ -33,8 +33,6 @@ export class xBullModule implements ModuleInterface {
       networkPassphrase?: string;
       address?: string;
       path?: string;
-      submit?: boolean;
-      submitUrl?: string;
     },
   ): Promise<{ signedTxXdr: string; signerAddress?: string }> {
     try {